add sweet alert

+++ wwwroot/js/games-index.js
@@ -1,19 +1,34 @@
 $(document).ready(function () {
     $('.js-delete').on('click', function () {
         var btn = $(this);
-        console.log(btn.data('id'));
-
-        $.ajax({
-            url: `/Games/Delete/${btn.data('id')}`,
-            method: 'DELETE',
-            success: function () {
-                alert('success');
-                btn.parents('tr').fadeOut(); 
-            },
-            error: function () {
-                alert('error');
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $.ajax({
+                    url: `/Games/Delete/${btn.data('id')}`,
+                    method: 'DELETE',
+                    success: function () {
+                        btn.parents('tr').fadeOut();
+                    },
+                    error: function () {
+                        alert('error');
+                    }
+                });
+                Swal.fire({
+                    title: "Deleted!",
+                    text: "Your file has been deleted.",
+                    icon: "success"
+                });
             }
         });
-        
+
+
     });
 });
